refactor: additional refactoring and bugfixes
