+++ Editor/ReactiveObjects/AnimationGeneration/ReactiveObjectPass.cs
@@ -33,13 +33,14 @@ namespace nadena.dev.modular_avatar.core.editor
 
         internal void Execute()
         {
-            Dictionary<TargetProp, AnimatedProperty> shapes = FindShapes(context);
-            FindObjectToggles(shapes, context);
-            FindMaterialSetters(shapes, context);
-
-            AnalyzeConstants(shapes);
-            ResolveToggleInitialStates(shapes);
-            PreprocessShapes(shapes, out var initialStates, out var deletedShapes);
+            Dictionary<TargetProp, AnimatedProperty> shapes =
+                new ReactiveObjectAnalyzer(context).Analyze(
+                    context.AvatarRootObject, 
+                    out var initialStates, 
+                    out var deletedShapes
+                );
+            
+            GenerateActiveSelfProxies(shapes);
             
             ProcessInitialStates(initialStates);
             ProcessInitialAnimatorVariables(shapes);
@@ -52,42 +53,19 @@ namespace nadena.dev.modular_avatar.core.editor
             ProcessMeshDeletion(deletedShapes);
         }
 
-        private void AnalyzeConstants(Dictionary<TargetProp, AnimatedProperty> shapes)
+        private void GenerateActiveSelfProxies(Dictionary<TargetProp, AnimatedProperty> shapes)
         {
             var asc = context.Extension<AnimationServicesContext>();
-            HashSet<GameObject> toggledObjects = new();
-
-            foreach (var targetProp in shapes.Keys)
-                if (targetProp is { TargetObject: GameObject go, PropertyName: "m_IsActive" })
-                    toggledObjects.Add(go);
-
-            foreach (var group in shapes.Values)
+            
+            foreach (var prop in shapes.Keys)
             {
-                foreach (var actionGroup in group.actionGroups)
+                if (prop.TargetObject is GameObject go && prop.PropertyName == "m_IsActive")
                 {
-                    foreach (var condition in actionGroup.ControllingConditions)
-                        if (condition.ReferenceObject != null && !toggledObjects.Contains(condition.ReferenceObject))
-                            condition.IsConstant = asc.AnimationDatabase.ClipsForPath(asc.PathMappings.GetObjectIdentifier(condition.ReferenceObject)).IsEmpty;
-
-                    var i = 0;
-                    // Remove redundant conditions
-                    actionGroup.ControllingConditions.RemoveAll(c => c.IsConstant && c.InitiallyActive && (i++ != 0));
+                    // Ensure a proxy exists for each object we're going to be toggling.
+                    // TODO: is this still needed?
+                    asc.GetActiveSelfProxy(go);
                 }
-
-                // Remove any action groups with always-off conditions
-                group.actionGroups.RemoveAll(agk =>
-                    agk.ControllingConditions.Any(c => !c.InitiallyActive && c.IsConstant));
-                
-                // Remove all action groups up until the last one where we're always on
-                var lastAlwaysOnGroup = group.actionGroups.FindLastIndex(ag => ag.IsConstantOn);
-                if (lastAlwaysOnGroup > 0)
-                    group.actionGroups.RemoveRange(0, lastAlwaysOnGroup - 1);
             }
-
-            // Remove shapes with no action groups
-            foreach (var kvp in shapes.ToList())
-                if (kvp.Value.actionGroups.Count == 0)
-                    shapes.Remove(kvp.Key);
         }
 
         private void ProcessInitialAnimatorVariables(Dictionary<TargetProp, AnimatedProperty> shapes)
@@ -100,121 +78,6 @@ namespace nadena.dev.modular_avatar.core.editor
 
                 if (!initialValues.ContainsKey(condition.Parameter))
                     initialValues[condition.Parameter] = condition.InitialValue;
-            }
-        }
-
-        private void PreprocessShapes(Dictionary<TargetProp, AnimatedProperty> shapes, out Dictionary<TargetProp, object> initialStates, out HashSet<TargetProp> deletedShapes)
-        {
-            // For each shapekey, determine 1) if we can just set an initial state and skip and 2) if we can delete the
-            // corresponding mesh. If we can't, delete ops are merged into the main list of operations.
-            
-            initialStates = new Dictionary<TargetProp, object>();
-            deletedShapes = new HashSet<TargetProp>();
-
-            foreach (var (key, info) in shapes.ToList())
-            {
-                if (info.actionGroups.Count == 0)
-                {
-                    // never active control; ignore it entirely
-                    shapes.Remove(key);
-                    continue;
-                }
-                
-                var deletions = info.actionGroups.Where(agk => agk.IsDelete).ToList();
-                if (deletions.Any(d => d.ControllingConditions.All(c => c.IsConstantActive)))
-                {
-                    // always deleted
-                    shapes.Remove(key);
-                    deletedShapes.Add(key);
-                    continue;
-                }
-                
-                // Move deleted shapes to the end of the list, so they override all Set actions
-                info.actionGroups = info.actionGroups.Where(agk => !agk.IsDelete).Concat(deletions).ToList();
-
-                var initialState = info.actionGroups.Where(agk => agk.InitiallyActive)
-                    .Select(agk => agk.Value)
-                    .Prepend(info.currentState) // use scene state if everything is disabled
-                    .Last();
-
-                initialStates[key] = initialState;
-                
-                // If we're now constant-on, we can skip animation generation
-                if (info.actionGroups[^1].IsConstant)
-                {
-                    shapes.Remove(key);
-                }
-            }
-        }
-
-        private void ResolveToggleInitialStates(Dictionary<TargetProp, AnimatedProperty> groups)
-        {
-            var asc = context.Extension<AnimationServicesContext>();
-            
-            Dictionary<string, bool> propStates = new Dictionary<string, bool>();
-            Dictionary<string, bool> nextPropStates = new Dictionary<string, bool>();
-            int loopLimit = 5;
-
-            bool unsettled = true;
-            while (unsettled && loopLimit-- > 0)
-            {
-                unsettled = false;
-
-                foreach (var group in groups.Values)
-                {
-                    if (group.TargetProp.PropertyName != "m_IsActive") continue;
-                    if (!(group.TargetProp.TargetObject is GameObject targetObject)) continue;
-
-                    var pathKey = asc.GetActiveSelfProxy(targetObject);
-
-                    bool state;
-                    if (!propStates.TryGetValue(pathKey, out state)) state = targetObject.activeSelf;
-
-                    foreach (var actionGroup in group.actionGroups)
-                    {
-                        bool evaluated = true;
-                        foreach (var condition in actionGroup.ControllingConditions)
-                        {
-                            if (!propStates.TryGetValue(condition.Parameter, out var propCondition))
-                            {
-                                propCondition = condition.InitiallyActive;
-                            }
-
-                            if (!propCondition)
-                            {
-                                evaluated = false;
-                                break;
-                            }
-                        }
-
-                        if (evaluated)
-                        {
-                            state = (float) actionGroup.Value > 0.5f;
-                        }
-                    }
-
-                    nextPropStates[pathKey] = state;
-
-                    if (!propStates.TryGetValue(pathKey, out var oldState) || oldState != state)
-                    {
-                        unsettled = true;
-                    }
-                }
-                
-                propStates = nextPropStates;
-                nextPropStates = new();
-            }
-
-            foreach (var group in groups.Values)
-            {
-                foreach (var action in group.actionGroups)
-                {
-                    foreach (var condition in action.ControllingConditions)
-                    {
-                        if (propStates.TryGetValue(condition.Parameter, out var state))
-                            condition.InitialValue = state ? 1.0f : 0.0f;
-                    }
-                }
             }
         }
 
@@ -281,7 +144,7 @@ namespace nadena.dev.modular_avatar.core.editor
                         switch (prop.propertyType)
                         {
                             case SerializedPropertyType.Boolean:
-                                animBaseState = prop.boolValue ? 1 : 0;
+                                animBaseState = prop.boolValue ? 1.0f : 0.0f;
                                 prop.boolValue = ((float)initialState) > 0.5f;
                                 break;
                             case SerializedPropertyType.Float:
@@ -508,7 +371,7 @@ namespace nadena.dev.modular_avatar.core.editor
             return asm;
         }
 
-        private AnimatorCondition[] GetTransitionConditions(AnimationServicesContext asc, ReactionData group)
+        private AnimatorCondition[] GetTransitionConditions(AnimationServicesContext asc, ReactionRule group)
         {
             var conditions = new List<AnimatorCondition>();
